+++ game/app/animator_controller.py
@@ -3,7 +3,7 @@ from typing import Any
 
 import pygame as pg
 
-from app import env
+from app import consts
 from app.base.animator import Animator, Animation
