+++ game/app/spells/water_blast.py
@@ -1,4 +1,4 @@
-from app import env
+from app import consts
 from app.base.animator import Animation, Animator
 
 import pygame as pg
